GUI-2/requirements.txt
Flask
numpy
pandas
pillow
ultralytics
joblib
scikit-learn
